+++ Sources/Lemmy-Swift-Client/Lemmy_Swift_Client.swift
@@ -119,18 +119,3 @@ public enum HTTPMethod: String {
 	case post = "POST"
 	case put = "PUT"
 }
-
-extension String {
-	func snakeCased() -> String? {
-		let pattern = "([a-z0-9])([A-Z])"
-
-		let regex = try? NSRegularExpression(pattern: pattern, options: [])
-		let range = NSRange(location: 0, length: count)
-		return regex?.stringByReplacingMatches(
-			in: self,
-			options: [],
-			range: range,
-			withTemplate: "$1_$2"
-		).lowercased()
-	}
-}
