+++ Sources/Lemmy-Swift-Client/String+snakeCased.swift
@@ -1,0 +1,23 @@
+//
+//  File.swift
+//  
+//
+//  Created by Colin Walsh on 12/8/23.
+//
+
+import Foundation
+
+extension String {
+    func snakeCased() -> String? {
+        let pattern = "([a-z0-9])([A-Z])"
+
+        let regex = try? NSRegularExpression(pattern: pattern, options: [])
+        let range = NSRange(location: 0, length: count)
+        return regex?.stringByReplacingMatches(
+            in: self,
+            options: [],
+            range: range,
+            withTemplate: "$1_$2"
+        ).lowercased()
+    }
+}
